# docs/swagger.yaml
basePath: /
definitions:
  model.Characteristic:
    properties:
      allowed_values:
        items: {}
        type: array
      display_unit:
        type: string
      id:
        type: string
      max_value: {}
      min_value: {}
      name:
        type: string
      sub_characteristics:
        items:
          $ref: '#/definitions/models.Characteristic'
        type: array
      type:
        $ref: '#/definitions/models.Type'
      value: {}
    type: object
  model.Criteria:
    properties:
      aspect_id:
        type: string
      device_class_id:
        type: string
      function_id:
        type: string
      interaction:
        $ref: '#/definitions/model.Interaction'
    type: object
  model.FilterPossibility:
    enum:
    - device
    - device_service_group
    - group
    - import
    type: string
    x-enum-varnames:
    - DeviceFilter
    - DeviceServiceGroupFilter
    - GroupFilter
    - ImportFilter
  model.Interaction:
    enum:
    - event
    - request
    - event+request
    type: string
    x-enum-varnames:
    - EVENT
    - REQUEST
    - EVENT_AND_REQUEST
  model.IotDescription:
    properties:
      criteria:
        items:
          $ref: '#/definitions/model.Criteria'
        type: array
      entity_only:
        type: boolean
      needs_same_entity_id_in_parameter:
        type: string
      type_filter:
        items:
          $ref: '#/definitions/model.FilterPossibility'
        type: array
    type: object
  model.MaintenanceProcedure:
    properties:
      bpmn_id:
        type: string
      internal_event_id:
        type: string
      message_ref:
        type: string
      parameter_descriptions:
        items:
          $ref: '#/definitions/model.ParameterDescription'
        type: array
      public_event_id:
        type: string
    type: object
...
  model.Option:
    properties:
      entity_id:
        description: to be used in combination with needs_same_entity_id_in_parameter
        type: string
      kind:
        description: optional helper for frontend to group options
        type: string
      label:
        type: string
      needs_same_entity_id_in_parameter:
        description: optional, frontend should filter this option depending on selection
          in referenced parameter
        type: string
      value: {}
    type: object
  model.ParameterDescription:
    properties:
      auto_select_all:
        type: boolean
      characteristic:
        $ref: '#/definitions/model.Characteristic'
      characteristic_id:
        type: string
      default_value: {}
      description:
        type: string
      id:
        type: string
      iot_description:
        $ref: '#/definitions/model.IotDescription'
      label:
        type: string
      multiple:
        type: boolean
      optional:
        type: boolean
      options:
        additionalProperties: true
        type: object
      order:
        type: integer
      type:
        type: string
    type: object
  model.PermissionsInfo:
    properties:
      permissions:
        additionalProperties:
          type: boolean
        type: object
      shared:
        type: boolean
    type: object
...
  model.SmartServiceExtendedParameter:
    properties:
      characteristic:
        $ref: '#/definitions/model.Characteristic'
      characteristic_id:
        type: string
      default_value: {}
      description:
        type: string
      has_no_valid_option:
        type: boolean
      id:
        type: string
      label:
        type: string
      multiple:
        description: 'if true: Value = new([]Type); if false: Value = new(Type);'
        type: boolean
      optional:
        type: boolean
      options:
        description: if null -> "free text/number/etc"
        items:
          $ref: '#/definitions/model.Option'
        type: array
      order:
        type: integer
      type:
        $ref: '#/definitions/model.Type'
      value: {}
      value_label:
        type: string
    type: object
  model.SmartServiceInstance:
    properties:
      created_at:
        description: unix timestamp, set by service on creation
        type: integer
      deleting:
        type: boolean
      description:
        type: string
      design_id:
        type: string
      error:
        description: is set if module-worker notifies the repository about a error
        type: string
      id:
        type: string
      name:
        type: string
      new_release_id:
        type: string
      parameters:
        items:
          $ref: '#/definitions/model.SmartServiceParameter'
        type: array
      ready:
        type: boolean
      release_id:
        type: string
      running_maintenance_ids:
        items:
          type: string
        type: array
      updated_at:
        description: unix timestamp, set by service on creation
        type: integer
      user_id:
        type: string
    type: object
...
  model.SmartServiceInstanceInit:
    properties:
      description:
        type: string
      name:
        type: string
      parameters:
        items:
          $ref: '#/definitions/model.SmartServiceParameter'
        type: array
    type: object
...
  model.SmartServiceParameter:
    properties:
      id:
        type: string
      label:
        type: string
      value: {}
      value_label:
        type: string
    type: object
  model.SmartServiceRelease:
    properties:
      created_at:
        description: unix timestamp, set by service on creation
        type: integer
      creator:
        type: string
      description:
        type: string
      design_id:
        type: string
      id:
        type: string
      name:
        type: string
      new_release_id:
        type: string
    type: object
  model.SmartServiceReleaseExtended:
    properties:
      bpmn_xml:
        type: string
      created_at:
        description: unix timestamp, set by service on creation
        type: integer
      creator:
        type: string
      description:
        type: string
      design_id:
        type: string
      id:
        type: string
      name:
        type: string
      new_release_id:
        type: string
      parsed_info:
        $ref: '#/definitions/model.SmartServiceReleaseInfo'
      permissions_info:
        allOf:
        - $ref: '#/definitions/model.PermissionsInfo'
        description: optional, set if query parameter permissions_info=true
      svg_xml:
        type: string
    type: object
  model.SmartServiceReleaseInfo:
    properties:
      maintenance_procedures:
        items:
          $ref: '#/definitions/model.MaintenanceProcedure'
        type: array
      parameter_descriptions:
        items:
          $ref: '#/definitions/model.ParameterDescription'
        type: array
    type: object
  model.Type:
    enum:
    - https://schema.org/Text
    - https://schema.org/Integer
    - https://schema.org/Float
    - https://schema.org/Boolean
    type: string
    x-enum-varnames:
    - String
    - Integer
    - Float
    - Boolean
  models.Characteristic:
    properties:
      allowed_values:
        items: {}
        type: array
      display_unit:
        type: string
      id:
        type: string
      max_value: {}
      min_value: {}
      name:
        type: string
      sub_characteristics:
        items:
          $ref: '#/definitions/models.Characteristic'
        type: array
      type:
        $ref: '#/definitions/models.Type'
      value: {}
    type: object
  models.Type:
    enum:
    - https://schema.org/Text
    - https://schema.org/Integer
    - https://schema.org/Float
    - https://schema.org/Boolean
    - https://schema.org/ItemList
    - https://schema.org/StructuredValue
    type: string
    x-enum-varnames:
    - String
    - Integer
    - Float
    - Boolean
    - List
    - Structure
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Smart-Service-Repository API
  version: "0.1"
paths:
...
  /extended-releases:
    get:
      description: returns a list of smart-service releases
      parameters:
      - description: limits size of result
        in: query
        name: limit
        type: integer
      - description: offset to be used in combination with limit
        in: query
        name: offset
        type: integer
      - description: 'rights needed to see a release; bay be a combination of the
          following letters: ''rwxa''; default = r; release rights are set with https://github.com/SENERGY-Platform/permission-command'
        in: query
        name: rights
        type: string
      - description: describes the sorting in the form of name.asc
        in: query
        name: sort
        type: string
      - description: optional text search (permission-search/elastic-search behavior)
        in: query
        name: search
        type: string
      - description: returns only newest release of the same design
        in: query
        name: latest
        type: boolean
      - description: add 'usable' flag to result, describing if the user hase options
          for all iot parameters
        in: query
        name: add-usable-flag
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Total-Count:
              description: count of all matching elements; used for pagination
              type: integer
          schema:
            items:
              $ref: '#/definitions/model.SmartServiceReleaseExtended'
            type: array
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: returns a list of smart-service releases
      tags:
      - releases
  /extended-releases/{id}:
    get:
      description: returns a smart-service release
      parameters:
      - description: Release ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SmartServiceReleaseExtended'
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: returns a smart-service release
      tags:
      - releases
...
  /releases:
    get:
      description: returns a list of smart-service releases
      parameters:
      - description: limits size of result
        in: query
        name: limit
        type: integer
      - description: offset to be used in combination with limit
        in: query
        name: offset
        type: integer
      - description: 'rights needed to see a release; bay be a combination of the
          following letters: ''rwxa''; default = r; release rights are set with https://github.com/SENERGY-Platform/permission-command'
        in: query
        name: rights
        type: string
      - description: describes the sorting in the form of name.asc
        in: query
        name: sort
        type: string
      - description: optional text search (permission-search/elastic-search behavior)
        in: query
        name: search
        type: string
      - description: returns only newest release of the same design
        in: query
        name: latest
        type: boolean
      - description: add 'usable' flag to result, describing if the user hase options
          for all iot parameters
        in: query
        name: add-usable-flag
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Total-Count:
              description: count of all matching elements; used for pagination
              type: integer
          schema:
            items:
              $ref: '#/definitions/model.SmartServiceRelease'
            type: array
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: returns a list of smart-service releases
      tags:
      - releases
    post:
      consumes:
      - application/json
      description: creates a smart-service release
      parameters:
      - description: Release ID
        in: path
        name: id
        required: true
        type: string
      - description: SmartServiceRelease
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/model.SmartServiceRelease'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SmartServiceRelease'
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: create a smart-service release
      tags:
      - releases
  /releases/{id}:
    delete:
      consumes:
      - application/json
      description: removes a smart-service release
      parameters:
      - description: Release ID
        in: path
        name: id
        required: true
        type: string
      - description: Delete all previous releases as well
        in: query
        name: delete_previous_releases
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
      summary: removes a smart-service release
      tags:
      - releases
    get:
      description: returns a smart-service release
      parameters:
      - description: Release ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SmartServiceRelease'
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: returns a smart-service release
      tags:
      - releases
  /releases/{id}/instances:
    post:
      consumes:
      - application/json
      description: creates a smart-service instance from the release
      parameters:
      - description: Release ID
        in: path
        name: id
        required: true
        type: string
      - description: SmartServiceInstanceInit
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/model.SmartServiceInstanceInit'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SmartServiceInstance'
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: creates a smart-service instance from the release
      tags:
      - releases
      - instances
  /releases/{id}/parameters:
    get:
      description: returns parameters of a release
      parameters:
      - description: Release ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.SmartServiceExtendedParameter'
            type: array
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: returns parameters of a release
      tags:
      - releases
      - parameter
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"